Re-adjusts vnorm calculation for better gameplay.
The vnorm cap that limits the speed of balls depending on their size
influences the game play. This commits re-adjusts that calculation
to make it better adjusted to the ball size.

The largest speed cap is chosen so that the largest ball bounces
slightly above the middle of the screen.

The lowest speed cap is chosen so that the smallest ball bounces high
enough for the player to be able to avoid it by sliding underneath. I
think this last aspect curcially affects how much fun players get to
have, as it is a skill that they can learn easily.