Implements (some weak form of) time reversal.
Adds the possibility of reversing time using the 'r' key during
gameplay.

The collision system does not work well when run backwards.
Surprisingly enough, it has nothing to do with switching, but with the
calculation of velocities, which are *not* negated before calculating
the resulting collision velocity. For this reason, a ball that is
bouncing back against the floor will be pushed even further underneath
it (but running forwards again recovers the correct position).